# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: OpenAccounting API
  version: "1.0.0"
  description: |
    REST API for the OpenAccounting accounting platform.

    ## Authentication
    All `/api/v1/...` endpoints require a Bearer token:
    ```
    Authorization: Bearer oa_live_…
    ```

    ## Idempotency
    `POST` endpoints accept an `Idempotency-Key` header. Replays
    within 24 hours return the original response without re-executing
    the mutation. Keys are scoped per token + ledger.

    ## Pagination
    List endpoints return `{"data": [...]}` with a `Link: <…>; rel="next"`
    header when more results are available. Pass `?limit=` (default 50,
    max 200) and `?cursor=` from the previous `Link` header.

    ## Errors
    All errors return RFC 7807 problem-details JSON
    (`Content-Type: application/problem+json`).

    ## Incoming Events
    External systems can push signed events to
    `POST /api/events/{ledger_id}` using the ledger's
    `incoming_events_secret` (HMAC-SHA256 in the
    `X-OA-Event-Signature: sha256=<hex>` header).

    ## Automation Rules
    Ledger owners create rules via the web UI at
    `/ledgers/{id}/automations`. Rules map an allowlisted trigger
    (event type) + conditions → an action (webhook / email /
    categorise transaction) and execute on the jobs queue.
servers:
  - url: /
    description: Same origin
paths:
...
  /ledgers:
    get:
      operationId: listLedgers
      summary: List ledgers owned by the caller
      tags: [Ledgers]
      responses:
        "200":
          description: Ledger list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LedgerList"
    post:
      operationId: createLedger
      summary: Create a new ledger
      tags: [Ledgers]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLedgerBody"
      responses:
        "201":
          description: Ledger created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ledger"
  /ledgers/{id}:
    get:
      operationId: getLedger
      summary: Get a single ledger
      tags: [Ledgers]
      parameters:
        - $ref: "#/components/parameters/LedgerIdPath"
      responses:
        "200":
          description: Ledger details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ledger"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateLedger
      summary: Update ledger name, timezone, or basis
      tags: [Ledgers]
      parameters:
        - $ref: "#/components/parameters/LedgerIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLedgerBody"
      responses:
        "200":
          description: Updated ledger
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ledger"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /ledgers/{ledger_id}/payments:
    get:
      operationId: listPayments
      summary: List payments in a ledger
      tags: [Payments]
      parameters:
        - $ref: "#/components/parameters/LedgerIdPath"
        - $ref: "#/components/parameters/PaginationParams"
      responses:
        "200":
          description: Payment list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentList"
    post:
      operationId: createPayment
      summary: Record a payment (idempotent)
      tags: [Payments]
      parameters:
        - $ref: "#/components/parameters/LedgerIdPath"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "201":
          description: Payment created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
  /ledgers/{ledger_id}/payments/{id}:
    get:
      operationId: getPayment
      summary: Get a single payment
      tags: [Payments]
      parameters:
        - $ref: "#/components/parameters/LedgerIdPath"
        - $ref: "#/components/parameters/PaymentIdPath"
      responses:
        "200":
          description: Payment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    LedgerIdPath:
      name: ledger_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PaymentIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
      description: Client-generated key for idempotent POSTs (24 h window)
    PaginationParams:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 200
          default: 50
      - name: cursor
        in: query
        required: false
        schema:
          type: string
  schemas:
    Ledger:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        base_currency:
          type: string
        timezone:
          type: string
        basis:
          type: string
        append_only:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    LedgerList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Ledger"
    CreateLedgerBody:
      type: object
      required: [name, base_currency]
      properties:
        name:
          type: string
        base_currency:
          type: string
        timezone:
          type: string
          default: UTC
        basis:
          type: string
          enum: [accrual, cash]
          default: accrual
    UpdateLedgerBody:
      type: object
      properties:
        name:
          type: string
        timezone:
          type: string
        basis:
          type: string
          enum: [accrual, cash]
...
    Payment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ledger_id:
          type: string
          format: uuid
        amount:
          type: number
        payment_date:
          type: string
          format: date
        payment_method:
          type: string
        kind:
          type: string
    PaymentList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
    CreatePaymentBody:
      type: object
      required: [amount, payment_date, payment_method, kind]
      properties:
        contact_id:
          type: string
          format: uuid
        invoice_id:
          type: string
          format: uuid
        amount:
          type: number
        payment_date:
          type: string
          format: date
        payment_method:
          type: string
          enum: [cash, check, bank_transfer, credit_card, other]
        reference:
          type: string
        kind:
          type: string
          enum: [received, made]
...